move  src/demo/h3-worldmap-demo.js -> dev/h3-worldmap-demo.js, src/demo/h3-worldmap-projection-select.js -> h3-worldmap-projection-select.js, update dev/index.html
