+++ dev/index.html
@@ -7,7 +7,7 @@
     <script src="../node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
     <script src="../node_modules/lit/polyfill-support.js"></script>
     <script type="module" src="../h3-worldmap.js"></script>
-    <script type="module" src="../src/demo/h3-worldmap-demo.js"></script>
+    <script type="module" src="../dev/h3-worldmap-demo.js"></script>
     <style>
       body { font-family: 'Georgia'; font-size: 1rem; }
       h3-worldmap {
